feat(viewer): Move tests to IFrameLoader
