+++ src/lib/viewers/iframe/__tests__/IFrameLoader-test.js
@@ -1,10 +1,23 @@
-/* eslint-disable no-unused-expressions */
 import IFrameLoader from '../IFrameLoader';
 import IFrameViewer from '../IFrameViewer';
 
 const sandbox = sinon.createSandbox();
 
 describe('lib/viewers/iframe/IFrameLoader', () => {
+    const iframe = new IFrameViewer({
+        file: {
+            id: '123',
+            extension: 'boxnote',
+            representations: {
+                entries: [
+                    {
+                        representation: 'ORIGINAL',
+                    },
+                ],
+            },
+        },
+    });
+
     afterEach(() => {
         sandbox.verifyAndRestore();
     });
@@ -18,4 +31,24 @@ describe('lib/viewers/iframe/IFrameLoader', () => {
             EXT: ['boxnote', 'boxdicom'],
         });
     });
+
+    test.each([
+        // disableDicom, fileType, viewerInstance
+        [true, 'boxdicom', undefined],
+        [true, 'boxnote', IFrameLoader.viewers[0]],
+        [false, 'boxdicom', IFrameLoader.viewers[0]],
+    ])(
+        'should return correct result depending on the disableDicom viewer option and file type',
+        (disableDicom, fileType, viewerInstance) => {
+            iframe.options.file.extension = fileType;
+
+            const viewerOptions = {
+                IFrame: {
+                    disableDicom,
+                },
+            };
+            const viewer = IFrameLoader.determineViewer(iframe.options.file, [], viewerOptions);
+            expect(viewer).toEqual(viewerInstance);
+        },
+    );
 });
